compose: Allow sending message to unsuscribed streams.
Previosly, sending message to unsuscribed streams was not allowed and an
error message is shown that the user is not suscribed.

Allow users to send messages to unsuscribed stream with the error
message still being shown.

Fixes: #21433
